refactor: 레이아웃과 라우터 분리
