+++ src/App.tsx
@@ -3,13 +3,16 @@ import { ThemeProvider } from 'styled-components';
 import GlobalStyles from './styles/GlobalStyles';
 import { theme } from './styles/theme';
 import Router from './router/Router';
+import Layout from './components/layout/Layout';
 
 function App() {
   return (
     <BrowserRouter>
       <ThemeProvider theme={theme}>
         <GlobalStyles />
-        <Router />
+        <Layout>
+          <Router />
+        </Layout>
       </ThemeProvider>
     </BrowserRouter>
   );
